requirements.txt
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
greenlet==3.1.1
idna==3.10
kombu==5.4.2
Mako==1.3.8
MarkupSafe==3.0.2
prompt_toolkit==3.0.50
psycopg2==2.9.10
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whatsapp-api-client-python==0.0.47
whatsapp-chatbot-python==0.9.5
